Move the list of icon providers to a separate class
